fix: /adminshop avec permissions admin
