+++ slash_commands/economy/shop.js
@@ -1,41 +1,26 @@
 const { MessageEmbed, MessageActionRow, MessageButton, MessageSelectMenu } = require('discord.js');
 const { MESSAGES } = require("../../util/constants");
 const { createError, sendLogs } = require("../../util/envoiMsg");
-const { YELLOW, NIGHT } = require("../../data/colors.json");
-const { CHECK_MARK, CROSS_MARK, NO_SUCCES } = require('../../data/emojis.json');
-const { sendMsgHubGroup, createReactionCollectorGroup, editMsgHubGroup, deleteMsgHubGroup } = require("../../util/msg/group");
-const { PREFIX, MONEY, CHANNEL } = require('../../config.js');
+const { YELLOW,  } = require("../../data/colors.json");
+const { CHECK_MARK, NO_SUCCES } = require('../../data/emojis.json');
+const { MONEY } = require('../../config.js');
 const moment = require('moment');
-const { GameItem } = require('../../models');
 const mongoose = require('mongoose');
 
 const NB_PAR_PAGES = 10;
 
 module.exports.run = async (interaction) => {
-    const group = interaction.options._group; // 'list', 'admin'
     const subcommand = interaction.options.getSubcommand();
 
-    if (group) {
-        if (group === 'admin') {
-            if (subcommand === 'cancel') {
-                cancel(interaction, interaction.options)
-            } else if (subcommand === 'refund') {
-                refund(interaction, interaction.options)
-            } else if (subcommand === 'delete') {
-                deleteItem(interaction, interaction.options)
-            }
-        }
-    } else {
-        if (subcommand === 'list') {
-            listGames(interaction, interaction.options);
-        } else if (subcommand === 'jeux') {
-            list(interaction, interaction.options, true);
-        } else if (subcommand === 'custom') {
-            list(interaction, interaction.options, false);
-            // TODO EN CONSTRUCTION
-        } else if (subcommand === 'sell') {
-            sell(interaction, interaction.options)
-        } 
+    if (subcommand === 'list') {
+        listGames(interaction, interaction.options);
+    } else if (subcommand === 'jeux') {
+        list(interaction, interaction.options, true);
+    } else if (subcommand === 'custom') {
+        list(interaction, interaction.options, false);
+        // TODO EN CONSTRUCTION
+    } else if (subcommand === 'sell') {
+        sell(interaction, interaction.options)
     }
 }
 
@@ -47,7 +32,7 @@ const list = async (interaction, options, showGame = false) => {
 
     let userDB = await client.getUser(author);
     if (!userDB)
-        return interaction.reply({ embeds: [createError(`Tu n'as pas de compte ! Merci de t'enregistrer avec la commande : \`${PREFIX}register\``)] });
+        return interaction.reply({ embeds: [createError(`Tu n'as pas de compte ! Merci de t'enregistrer avec la commande : \`/register\``)] });
 
     let infos = {};
     infos.money = userDB.money;
@@ -416,7 +401,7 @@ async function listGames(interaction, options) {
 
     let userDB = await client.getUser(author);
     if (!userDB)
-        return interaction.reply({ embeds: [createError(`${author.user.tag} n'a pas encore de compte ! Pour s'enregistrer : \`${PREFIX}register\``)] });
+        return interaction.reply({ embeds: [createError(`${author.user.tag} n'a pas encore de compte ! Pour s'enregistrer : \`/register\``)] });
     
     const items = await client.findGameItemShopByGame();
     let embed = new MessageEmbed()
@@ -534,7 +519,7 @@ async function sell(interaction, options) {
 
     let userDB = await client.getUser(author);
     if (!userDB)
-        return interaction.reply({ embeds: [createError(`${author.user.tag} n'a pas encore de compte ! Pour s'enregistrer : \`${PREFIX}register\``)] });
+        return interaction.reply({ embeds: [createError(`${author.user.tag} n'a pas encore de compte ! Pour s'enregistrer : \`/register\``)] });
 
     if (montant < 0)
         return interaction.reply({ embeds: [createError(`Montant négatif !`)] });
@@ -621,137 +606,6 @@ async function sell(interaction, options) {
 
     // envoie log 'Nouvel vente par @ sur jeu X' (voir avec Tobi)
     sendLogs(client, `Nouveau jeu dans le shop`, `${author} vient d'ajouter **${game.name}** à **${montant} ${MONEY}** !`, `ID : ${itemDB._id}`, YELLOW);
-}
-
-/* ADMIN */
-async function cancel(interaction, options) {
-    const id = options.get('id')?.value;
-    const client = interaction.client;
-    const author = interaction.member;
-    
-    let gameItem;
-    try {
-        gameItem = await client.findGameItemShop({ _id: id });
-        if (gameItem.length === 0) throw `Non trouvée`;
-    } catch (error) {
-        return interaction.reply({ embeds: [createError(`Vente non trouvée`)] });
-    }
-    // on recup [0] car findGameItemShop retourne un array..
-    gameItem = gameItem[0];
-
-    // teste si state existe et si != 'done'
-    if (!gameItem.state) return interaction.reply({ embeds: [createError(`La vente n'a pas encore **commencée** ! Utiliser \`/shop admin delete <id>\``)] });
-    if (gameItem.state === 'done') return interaction.reply({ embeds: [createError(`La vente est déjà **terminée** ! Utiliser \`/shop admin refund <id>\``)] });
-
-    await client.update(gameItem, { $unset : { state : 1} } )
-    logger.info(`Annulation vente id ${id}`);
-    // TODO embed plutot qu'emoji (car on peut pas juste react je crois)
-    interaction.reply(CHECK_MARK)
-    sendLogs(client, `Annulation vente`, `${author} a annulé la vente en cours de **${gameItem.game.name}**, par **${gameItem.seller.username}**`, `ID : ${id}`, YELLOW);
-}
-
-async function refund(interaction, options) {
-    const id = options.get('id')?.value;
-    const client = interaction.client;
-    const author = interaction.member;
-    
-    let gameItem;
-    try {
-        gameItem = await client.findGameItemShop({ _id: id });
-        if (gameItem.length === 0) throw `Non trouvée`;
-    } catch (error) {
-        return interaction.reply({ embeds: [createError(`Vente non trouvée`)] });
-    }
-    // on recup [0] car findGameItemShop retourne un array..
-    gameItem = gameItem[0];
-
-    // teste si state existe et si == 'done'
-    if (!gameItem.state) return interaction.reply({ embeds: [createError(`La vente n'a pas encore **commencée** ! Utiliser \`/shop admin delete <id>\``)] });
-    if (gameItem.state !== 'done') return interaction.reply({ embeds: [createError(`La vente n'est pas encore **terminée** ! Utiliser \`/shop admin cancel <id>\``)] });
-
-    // maj statut item
-    await client.update(gameItem, { $unset : { state : 1, buyer: 1 } } )
-
-    // rembourse acheteur
-    try {
-        await client.update(gameItem.buyer, { money: gameItem.buyer.money + gameItem.montant })
-    } catch (error) {
-        return interaction.reply({ embeds: [createError(`Acheteur non trouvé ! Impossible de le rembourser.`)] });
-    }
-
-    // reprend argent au vendeur
-    try {
-        await client.update(gameItem.seller, { money: gameItem.seller.money - gameItem.montant })
-    } catch (error) {
-        return interaction.reply({ embeds: [createError(`Vendeur non trouvé ! Impossible de lui reprendre l'argent.`)] });
-    }
-
-    logger.info(`Remboursement vente id ${id}`);
-    // TODO embed plutot qu'emoji (car on peut pas juste react je crois)
-    interaction.reply(CHECK_MARK)
-    sendLogs(client, `Annulation vente`, `${author} a annulé la vente, pour rembourser l'achat de **${gameItem.buyer.username}**, du jeu **${gameItem.game.name}**, vendu par **${gameItem.seller.username}**`, `ID : ${id}`, YELLOW);
-}
-
-async function deleteItem(interaction, options) {
-    // const id = options.get('id')?.value;
-    const jeu = options.get('jeu')?.value;
-    const vendeur = options.get('vendeur')?.user;
-    const client = interaction.client;
-    const author = interaction.member;
-    
-    // recup list item (pas encore vendu) en fonction vendeur ET jeu
-    let items = await client.findGameItemShopBy({ game: jeu, seller: vendeur.id, notSold: true });
-    items = items.map(i => ({ label: `${i.game.name}, vendu par ${i.seller.username}, à ${i.montant} ${MONEY}`, value: '' + i._id}))
-
-    // row contenant le Select menu
-    const row = new MessageActionRow().addComponents(
-        new MessageSelectMenu()
-            .setCustomId('select-items-' + author)
-            .setPlaceholder(`Sélectionner l'item à supprimer..`)
-            .addOptions(items)
-    );
-    let embed = new MessageEmbed()
-        .setColor(NIGHT)
-        .setTitle(`Quel item à supprimer ?`);
-    
-    let msgEmbed = await interaction.reply({embeds: [embed], components: [row], fetchReply: true});
-
-    // TODO choix select : delete
-    let filter, itr;
-    try {
-        filter = i => {return i.user.id === author.id}
-        itr = await msgEmbed.awaitMessageComponent({
-            filter,
-            componentType: 'SELECT_MENU',
-            time: 30000 // 5min
-        });
-    } catch (error) {
-        msgEmbed.edit({ components: [] })
-        return;
-    }
-    // on enleve le select
-    await interaction.editReply({ components: [] })
-
-    const id = itr.values[0]
-    const gameItem = await client.findGameItemShop({ _id: new mongoose.Types.ObjectId(id) });
-    logger.info(`.. Item ${gameItem[0]._id} choisi`);
-
-    // teste si state n'existe pas
-    if (gameItem[0].state) return interaction.editReply({ embeds: [createError(`L'item ne doit pas être encore en cours de vente ! Utiliser \`/shop admin cancel <id>\` ou \`/shop admin refund <id>\``)] });
-   
-    // suppr item boutique
-    try {
-        await client.deleteGameItemById(id);
-    } catch (error) {
-        return interaction.editReply({ embeds: [createError(`Item du shop non trouvé !`)] });
-    }
-
-    logger.info(`Suppression vente id ${id}`);
-    // TODO embed plutot qu'emoji (car on peut pas juste react je crois)
-    embed.setTitle(`${CHECK_MARK} Item supprimé`)
-    await interaction.editReply({ embeds: [embed] })
-    sendLogs(client, `Suppression vente`, `${author} a supprimé la vente de **${jeu}**, par **${vendeur}**`, `ID : ${id}`, YELLOW);
-    return;    
 }
 
 module.exports.help = MESSAGES.COMMANDS.ECONOMY.SHOP;
